+++ fwdpp/sampling_functions.hpp
@@ -16,9 +16,7 @@
   This collection of functions allows a user to draw a sample of size \f$n \ll 2N\f$ from
   a simulated population.
 
-  The library provides several overloads of the functions KTfwd::ms_sample and KTfwd::ms_sample_separate.  Further,
-  the behavior of these functions differs slightly depending on whether or not the output from the simulation 
-  comes from a gamete- or an indvidual-based method.
+  The library provides several overloads of the functions KTfwd::ms_sample and KTfwd::ms_sample_separate.
 
   The following features are in common to all versions of these functions:
 
@@ -82,8 +80,7 @@
   @defgroup samplingPopsInd Randomly-sampling diploids.
   @ingroup samplingPops
 
-  These functions pull a sample of \f$n\f$ simulated individuals from the population.  You may only use them on the 
-  output of individual-based simulations (gamete-based simulations never explicitly store individuals).
+  These functions pull a sample of \f$n\f$ simulated individuals from the population.
 
   The individuals are sampled uniformly, and with replacement, with no regard for their fitness.
 
@@ -121,35 +118,6 @@ namespace KTfwd
   std::vector<unsigned> sample_sfs(gsl_rng * r, 
 				   const container_t<gamete_type,allocator_t > & gametes,
 				   const unsigned & n, const unsigned & N);
-  /*!
-    Take a sample of size n from a larger population of N diploids
-    \return A vector  of variable sites. The first block corresponds to neutral variants, and the second to non-neutral
-    \ingroup samplingPopsGamete
-  */
-  template< typename gamete_type,
-	    typename allocator_t,
-	    template<typename,typename> class container_t>
-  typename std::enable_if< std::is_base_of<mutation_base,typename gamete_type::mutation_type>::value,
-			   std::vector< std::pair<double, std::string> > >::type
-  ms_sample(gsl_rng * r,
-	    const container_t<gamete_type,allocator_t > * gametes,
-	    const unsigned & n, const unsigned & N,
-	    bool remove_fixed = true);
-  
-  
-  /*!
-    \ingroup samplingPopsGamete
-  */
-  template< typename gamete_type,
-	    typename allocator_t,
-	    template<typename,typename> class container_t>
-  typename std::enable_if< std::is_base_of<mutation_base,typename gamete_type::mutation_type>::value,
-			   std::pair< std::vector< std::pair<double, std::string> > ,
-				      std::vector< std::pair<double, std::string> > > >::type
-  ms_sample_separate(gsl_rng * r,
-		     const container_t<gamete_type,allocator_t > * gametes,
-		     const unsigned & n, const unsigned & N,
-		     bool remove_fixed = true);
 
   /*!
     \brief Sampling from a population in an individual-based simulation